Move enums (not related to curl) to enums header
